Update example based on review
